fix(core): move to_matrix test out of georef.rs to clear the module-size ratchet
The to_matrix unit test added in this PR pushed georef.rs to 851 lines,
past its 793-line ratchet budget. Per the ratchet's own instruction
("shrink or split instead of raising the budget"), move the test rather
than editing the allowlist.

to_matrix and GeoReference are both pub, so the test now runs as an
ordinary integration test in rust/core/tests/georef_matrix.rs, matching
the existing convention for this crate's tests. georef.rs is back to
793 lines; the total ifc-lite-core test count (138) is unchanged.

Test-only change; no changeset needed.
